Address pages: add epoch data (#879)
