+++ apps/block_scout_web/lib/block_scout_web/controllers/address_coin_balance_controller.ex
@@ -10,6 +10,7 @@ defmodule BlockScoutWeb.AddressCoinBalanceController do
   import BlockScoutWeb.Models.GetAddressTags, only: [get_address_tags: 2]
 
   alias BlockScoutWeb.{AccessHelpers, AddressCoinBalanceView, Controller}
+  alias Explorer.Celo.EpochUtil
   alias Explorer.{Chain, Market}
   alias Explorer.Chain.{Address, Wei}
   alias Explorer.ExchangeRates.Token
@@ -79,7 +80,8 @@ defmodule BlockScoutWeb.AddressCoinBalanceController do
         exchange_rate: Market.get_exchange_rate(Explorer.coin()) || Token.null(),
         current_path: Controller.current_full_path(conn),
         counters_path: address_path(conn, :address_counters, %{"id" => Address.checksum(address_hash)}),
-        tags: get_address_tags(address_hash, current_user(conn))
+        tags: get_address_tags(address_hash, current_user(conn)),
+        celo_epoch: EpochUtil.get_address_summary(address)
       )
     else
       {:restricted_access, _} ->
@@ -105,7 +107,8 @@ defmodule BlockScoutWeb.AddressCoinBalanceController do
               exchange_rate: Market.get_exchange_rate(Explorer.coin()) || Token.null(),
               counters_path: address_path(conn, :address_counters, %{"id" => Address.checksum(address_hash)}),
               current_path: Controller.current_full_path(conn),
-              tags: get_address_tags(address_hash, current_user(conn))
+              tags: get_address_tags(address_hash, current_user(conn)),
+              celo_epoch: EpochUtil.get_address_summary(address)
             )
 
           _ ->
