+++ apps/block_scout_web/lib/block_scout_web/controllers/address_transaction_controller.ex
@@ -12,6 +12,7 @@ defmodule BlockScoutWeb.AddressTransactionController do
   import BlockScoutWeb.Models.GetAddressTags, only: [get_address_tags: 2]
 
   alias BlockScoutWeb.{AccessHelpers, Controller, TransactionView}
+  alias Explorer.Celo.EpochUtil
   alias Explorer.{Chain, Market}
   alias Explorer.Export.CSV
 
@@ -126,7 +127,8 @@ defmodule BlockScoutWeb.AddressTransactionController do
         filter: params["filter"],
         counters_path: address_path(conn, :address_counters, %{"id" => address_hash_string}),
         current_path: Controller.current_full_path(conn),
-        tags: get_address_tags(address_hash, current_user(conn))
+        tags: get_address_tags(address_hash, current_user(conn)),
+        celo_epoch: EpochUtil.get_address_summary(address)
       )
     else
       :error ->
@@ -156,7 +158,8 @@ defmodule BlockScoutWeb.AddressTransactionController do
               filter: params["filter"],
               counters_path: address_path(conn, :address_counters, %{"id" => address_hash_string}),
               current_path: Controller.current_full_path(conn),
-              tags: get_address_tags(address_hash, current_user(conn))
+              tags: get_address_tags(address_hash, current_user(conn)),
+              celo_epoch: EpochUtil.get_address_summary(address)
             )
 
           _ ->
