+++ apps/block_scout_web/lib/block_scout_web/controllers/address_write_contract_controller.ex
@@ -13,6 +13,7 @@ defmodule BlockScoutWeb.AddressWriteContractController do
 
   alias BlockScoutWeb.AccessHelpers
   alias BlockScoutWeb.AddressView
+  alias Explorer.Celo.EpochUtil
   alias Explorer.{Chain, Market}
   alias Explorer.Chain.Address
   alias Explorer.ExchangeRates.Token
@@ -51,7 +52,8 @@ defmodule BlockScoutWeb.AddressWriteContractController do
             non_custom_abi: true,
             coin_balance_status: CoinBalanceOnDemand.trigger_fetch(address),
             counters_path: address_path(conn, :address_counters, %{"id" => Address.checksum(address_hash)}),
-            tags: get_address_tags(address_hash, current_user(conn))
+            tags: get_address_tags(address_hash, current_user(conn)),
+            celo_epoch: EpochUtil.get_address_summary(address)
           ]
       )
     else
@@ -69,7 +71,8 @@ defmodule BlockScoutWeb.AddressWriteContractController do
                   non_custom_abi: false,
                   coin_balance_status: CoinBalanceOnDemand.trigger_fetch(address),
                   counters_path: address_path(conn, :address_counters, %{"id" => Address.checksum(address_hash)}),
-                  tags: get_address_tags(address_hash, current_user(conn))
+                  tags: get_address_tags(address_hash, current_user(conn)),
+                  celo_epoch: EpochUtil.get_address_summary(address)
                 ]
             )
           else
