+++ apps/block_scout_web/lib/block_scout_web/templates/address_signed/index.html.eex
@@ -7,7 +7,8 @@
     exchange_rate: @exchange_rate,
     coin_balance_status: @coin_balance_status,
     counters_path: @counters_path,
-    tags: @tags
+    tags: @tags,
+    celo_epoch: @celo_epoch
   %>
 
   <section data-page="blocks-signed">
